int equals fix
